+++ openhivenpy/types/member.py
@@ -1,3 +1,31 @@
+"""
+Member File which implements the Hiven Member type and its methods
+(endpoints)
+
+---
+
+Under MIT License
+
+Copyright © 2020 - 2021 Luna Klatzer
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
 # Used for type hinting and not having to use annotations for the objects
 from __future__ import annotations
 
@@ -95,18 +123,27 @@ class Member(User):
 
     @property
     def id(self) -> Optional[str]:
-        return getattr(self, '_user_id', None)
+        """ User-ID of the member. Alias for user_id """
+        return self.user_id
 
     @property
     def user_id(self) -> Optional[str]:
+        """ User-ID of the member """
         return getattr(self, '_user_id', None)
 
     @property
     def joined_house_at(self) -> Optional[str]:
+        """ Joined at date (unix-timestamp probably). Alias for joined_at """
+        return self.joined_at
+
+    @property
+    def joined_at(self) -> Optional[str]:
+        """ Joined at date (unix-timestamp probably) """
         return getattr(self, '_joined_at', None)
 
     @property
     def house(self) -> Optional[House]:
+        """ Parent House object instance of this class """
         from . import House
         if type(self._house) is str:
             house_id = self._house
@@ -130,15 +167,13 @@ class Member(User):
 
     @property
     def house_id(self) -> Optional[str]:
+        """ Returns the ID of the parent House """
         return getattr(self, '_house_id', None)
 
     @property
     def roles(self) -> Optional[List[dict]]:
+        """ Returns the roles of the Member """
         return getattr(self, '_roles', None)
-
-    @property
-    def joined_at(self) -> Optional[str]:
-        return getattr(self, '_joined_at', None)
 
     async def kick(self) -> bool:
         """