+++ openhivenpy/types/message.py
@@ -1,3 +1,31 @@
+"""
+Message File which implements the Hiven Message type and its methods
+(endpoints)
+
+---
+
+Under MIT License
+
+Copyright © 2020 - 2021 Luna Klatzer
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
 # Used for type hinting and not having to use annotations for the objects
 from __future__ import annotations
 
@@ -93,7 +121,8 @@ class Message(DataClassObject):
         self._timestamp = data.get('timestamp')
         self._edited_at = data.get('edited_at')
         self._mentions = data.get('mentions')
-        self._type = data.get('type')  # I believe, 0 = normal message, 1 = system.
+        # I believe, 0 = normal message, 1 = system.
+        self._type = data.get('type')
         self._exploding = data.get('exploding')
         self._house_id = data.get('house_id')
         self._house = data.get('house')
@@ -156,7 +185,9 @@ class Message(DataClassObject):
                 room_ = None
 
             if room_ is None:
-                raise InvalidPassedDataError("The passed room is not in the correct format!", data=data)
+                raise InvalidPassedDataError(
+                    "The passed room is not in the correct format!", data=data
+                )
             else:
                 data['room_id'] = room_
 
@@ -185,26 +216,34 @@ class Message(DataClassObject):
                 author = None
 
             if author is None:
-                raise InvalidPassedDataError("The passed author is not in the correct format!", data=data)
+                raise InvalidPassedDataError(
+                    "The passed author is not in the correct format!",
+                    data=data
+                )
             else:
                 data['author'] = author
 
         data['author'] = data['author_id']
         data['house'] = data['house_id']
         data['room'] = data['room_id']
-        data['device_id'] = utils.safe_convert(str, data.get('device_id'), None)
+        data['device_id'] = utils.safe_convert(
+            str, data.get('device_id'), None
+        )
         return data
 
     @property
     def id(self) -> Optional[str]:
+        """ ID of the message """
         return getattr(self, '_id', None)
 
     @property
     def author_id(self) -> Optional[str]:
+        """ ID of the parent Author """
         return getattr(self, '_author_id', None)
 
     @property
     def author(self) -> Optional[User]:
+        """ Returns the Author parent object instance """
         if type(self._author) is str:
             author_id = self._author
         elif type(self.author_id) is str:
@@ -226,14 +265,13 @@ class Message(DataClassObject):
             return None
 
     @property
-    def created_at(self) -> Optional[str]:
-        return getattr(self, '_created_at', None)
-
-    @property
     def timestamp(self) -> Optional[datetime.datetime]:
+        """ Returns the date the message was created (unix-timestamp) """
         if utils.convertible(int, self._timestamp):
             # Converting to seconds because it's in milliseconds
-            self._timestamp = datetime.datetime.fromtimestamp(utils.safe_convert(int, self._timestamp) / 1000)
+            self._timestamp = datetime.datetime.fromtimestamp(
+                utils.safe_convert(int, self._timestamp) / 1000
+            )
             return self._timestamp
         elif type(self._timestamp) is datetime.datetime:
             return self._timestamp
@@ -242,18 +280,22 @@ class Message(DataClassObject):
 
     @property
     def type(self) -> Optional[int]:
+        """ Returns the type of the message """
         return getattr(self, '_type', None)
 
     @property
     def exploding(self) -> Optional[bool]:
+        """ Returns whether the message is exploding """
         return getattr(self, '_exploding', None)
 
     @property
     def edited_at(self) -> Optional[str]:
+        """ Returns the date the message was edited (unix-timestamp) """
         return getattr(self, '_edited_at', None)
 
     @property
     def room(self) -> Optional[TextRoom]:
+        """ Returns the Room parent object the message was sent in """
         from . import TextRoom
         if type(self._room) is str:
             data = self._client.storage['rooms']['house'].get(self._room)
@@ -270,6 +312,9 @@ class Message(DataClassObject):
 
     @property
     def house(self) -> Optional[House]:
+        """
+        Returns the House parent object, if the message was sent inside a House
+        """
         if type(self._house) is str:
             data = self._client.storage['houses'].get(self._house)
             if data:
@@ -285,8 +330,11 @@ class Message(DataClassObject):
 
     @property
     def attachment(self) -> Optional[Attachment]:
+        """ Returns the Attachment of the message, if it has one """
         if type(self._attachment) is dict:
-            self._attachment: Union[Attachment, dict] = Attachment(data=self._attachment, client=self._client)
+            self._attachment: Union[Attachment, dict] = Attachment(
+                data=self._attachment, client=self._client
+            )
             return self._attachment
         elif type(self._attachment) is Attachment:
             return self._attachment
@@ -295,10 +343,12 @@ class Message(DataClassObject):
 
     @property
     def content(self) -> Optional[str]:
+        """ Returns the string content of the message """
         return getattr(self, '_content', None)
 
     @property
     def mentions(self) -> Optional[List[Mention]]:
+        """ Returns the mentions of the message """
         from . import Mention
         if type(self._mentions) is list:
             if type(self._mentions[0]) is dict:
@@ -319,31 +369,46 @@ class Message(DataClassObject):
 
     @property
     def room_id(self) -> Optional[str]:
+        """ Returns the id of the Room parent object """
         return getattr(self, '_room_id', None)
 
     @property
     def house_id(self) -> Optional[str]:
+        """
+        Returns the id of the House parent object,
+        if the message was sent inside a House
+        """
         return getattr(self, '_house_id', None)
 
     @property
+    def is_house_message(self) -> bool:
+        """ Returns whether the message was sent inside a House """
+        return self.house_id is not None
+
+    @property
     def embed(self) -> Embed:
+        """ Returns the Embed of the message, if it has one """
         return getattr(self, '_embed', None)
 
     @property
     def bucket(self) -> Optional[int]:
+        """ Returns the bucket of the message """
         return getattr(self, '_bucket', None)
 
     @property
     def device_id(self) -> Optional[str]:
+        """ Returns the device id of the author of the message """
         return getattr(self, '_device_id', None)
 
     @property
     def exploding_age(self) -> Optional[int]:
+        """ Returns the exploding age of the message """
         return getattr(self, '_exploding_age', None)
 
     async def mark_as_read(self, delay: float = None) -> bool:
         """
-        Marks the message as read. This doesn't need to be done for bot clients.
+        Marks the message as read. This doesn't need to be done for bot
+        clients.
         
         :param delay: Delay until marking the message as read (in seconds)
         :return: True if the request was successful else False
@@ -398,7 +463,8 @@ class Message(DataClassObject):
         """
         try:
             await self._client.http.patch(
-                endpoint=f"/rooms/{self.room_id}/messages/{self.id}", json={'content': content}
+                endpoint=f"/rooms/{self.room_id}/messages/{self.id}",
+                json={'content': content}
             )
             return True