+++ openhivenpy/types/private_room.py
@@ -1,3 +1,31 @@
+"""
+PrivateRoom File which implements the Hiven GroupPrivateRoom and PrivateRoom
+type and its methods (endpoints)
+
+---
+
+Under MIT License
+
+Copyright © 2020 - 2021 Luna Klatzer
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
 # Used for type hinting and not having to use annotations for the objects
 from __future__ import annotations
 
@@ -49,7 +77,9 @@ class PrivateGroupRoom(DataClassObject):
             ('recipients', self.recipients),
             ('type', self.type)
         ]
-        return '<PrivateGroupRoom {}>'.format(' '.join('%s=%s' % t for t in info))
+        return '<PrivateGroupRoom {}>'.format(
+            ' '.join('%s=%s' % t for t in info)
+        )
 
     def get_cached_data(self) -> Optional[dict]:
         """
@@ -76,17 +106,36 @@ class PrivateGroupRoom(DataClassObject):
         data = cls.validate(data)
         data['name'] = f"Private chat with {data['recipients'][0]['name']}"
 
-        if type(data.get('recipients')) is list:
-            data['recipients'] = [i['id'] for i in data['recipients']]
-
+        rep = data.get('recipients')
+        id_list: List[str] = []
+        if type(rep) is list:
+            for user in rep:
+                if type(user) is dict:
+                    id_list += str(user.get('id', None))
+                elif isinstance(user, DataClassObject):
+                    id_list += str(getattr(user, 'id', None))
+                else:
+                    raise InvalidPassedDataError(
+                        "The passed recipient is not in the correct "
+                        "format!",
+                        data=data
+                    )
+        else:
+            raise InvalidPassedDataError(
+                "The passed recipients are not in the correct format!",
+                data=data
+            )
+        data['recipients'] = id_list
         return data
 
     @property
     def client_user(self) -> Optional[User]:
+        """ Returns the Client User inside this PrivateGroupRoom """
         return getattr(self, '_client_user', None)
 
     @property
     def recipients(self) -> Optional[List[User]]:
+        """ Returns a list of all recipients """
         from . import User
         if utils.convertible(int, self._recipients):
             recipients = []
@@ -107,29 +156,38 @@ class PrivateGroupRoom(DataClassObject):
 
     @property
     def id(self) -> Optional[str]:
+        """ Returns the id of the PrivateGroupRoom """
         return getattr(self, '_id', None)
 
     @property
     def last_message_id(self) -> Optional[str]:
+        """ Returns the id of the last message inside the PrivateGroupRoom """
         return getattr(self, '_last_message_id', None)
 
     @property
     def name(self) -> Optional[str]:
+        """ Returns the name of the PrivateGroupRoom """
         return getattr(self, '_name', None)
 
     @property
     def description(self) -> Optional[int]:
+        """ Returns the description of the PrivateGroupRoom """
         return getattr(self, '_description', None)
 
     @property
     def emoji(self) -> Optional[str]:
+        """ Returns the emoji of this PrivateGroupRoom if it exists """
         return getattr(self, '_emoji', None)
 
     @property
     def type(self) -> Optional[int]:
+        """ Returns the type of this PrivateGroupRoom """
         return getattr(self, '_type', None)
 
-    async def send(self, content: str, delay: float = None) -> Optional[Message]:
+    # TODO! Implement Base Room class (ABC)
+    async def send(
+            self, content: str, delay: float = None
+    ) -> Optional[Message]:
         """
         Sends a message in the private room.
 
@@ -250,7 +308,10 @@ class PrivateRoom(DataClassObject):
                 name = None
 
             if recipient is None:
-                raise InvalidPassedDataError("The passed recipient/s is/are not in the correct format!", data=data)
+                raise InvalidPassedDataError(
+                    "The passed recipient/s is/are not in the correct format!",
+                    data=data
+                )
             else:
                 data['recipient_id'] = recipient
 
@@ -266,6 +327,7 @@ class PrivateRoom(DataClassObject):
 
     @property
     def name(self) -> Optional[str]:
+        """ Name of the PrivateRoom """
         if type(self._name) is None:
             self._name = f"Private chat with {self.recipient.name}"
             return self._name
@@ -276,14 +338,17 @@ class PrivateRoom(DataClassObject):
 
     @property
     def client_user(self) -> Optional[User]:
+        """ Returns the client_user of this class """
         return getattr(self, '_client_user', None)
 
     @property
     def recipient_id(self) -> Optional[str]:
+        """ The ID of the recipient """
         return getattr(self, '_recipient_id', None)
 
     @property
     def recipient(self) -> Optional[User]:
+        """ Returns the recipient object instance """
         from . import User
         if type(self._recipient) is str:
             recipient_id = self._recipient
@@ -307,22 +372,27 @@ class PrivateRoom(DataClassObject):
 
     @property
     def id(self) -> Optional[str]:
+        """ Returns the id of the PrivateRoom """
         return getattr(self, '_id', None)
 
     @property
     def description(self) -> Optional[str]:
+        """ Return the description of the PrivateRoom """
         return getattr(self, '_description', None)
 
     @property
     def emoji(self) -> Optional[str]:
+        """ The emoji of the PrivateRoom, if it has one """
         return getattr(self, '_emoji', None)
 
     @property
     def last_message_id(self) -> Optional[str]:
+        """ The id of the last sent message """
         return getattr(self, '_last_message_id', None)
 
     @property
     def type(self) -> Optional[int]:
+        """ The type of the PrivateRoom """
         return getattr(self, '_type', None)
 
     async def start_call(self, delay: float = None) -> bool:
@@ -353,6 +423,7 @@ class PrivateRoom(DataClassObject):
             # TODO! Raise exception
             return False
 
+    # TODO! Implement Base Room class (ABC)
     async def send(
             self, content: str, delay: float = None
     ) -> Optional[Message]: