+++ openhivenpy/types/usertyping.py
@@ -1,3 +1,29 @@
+"""
+UserTyping File which implements the UserTyping
+---
+
+Under MIT License
+
+Copyright © 2020 - 2021 Luna Klatzer
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
 # Used for type hinting and not having to use annotations for the objects
 from __future__ import annotations
 
@@ -46,9 +72,12 @@ class UserTyping(DataClassObject):
 
     @property
     def timestamp(self) -> Optional[datetime.datetime]:
+        """ Time-stamp of the User-Typing (unix) """
         if utils.convertible(int, self._timestamp):
             # Converting to seconds because it's in milliseconds
-            self._timestamp = datetime.datetime.fromtimestamp(utils.safe_convert(int, self._timestamp) / 1000)
+            self._timestamp = datetime.datetime.fromtimestamp(
+                utils.safe_convert(int, self._timestamp) / 1000
+            )
             return self._timestamp
         elif type(self._timestamp) is datetime.datetime:
             return self._timestamp
@@ -57,6 +86,7 @@ class UserTyping(DataClassObject):
 
     @property
     def author(self) -> Optional[User]:
+        """ Author object of the User-Typing Class """
         if type(self._author) is str:
             data = self._client.storage['users'].get(self._author)
             if data:
@@ -72,6 +102,7 @@ class UserTyping(DataClassObject):
 
     @property
     def house(self) -> Optional[House]:
+        """ House object of the Context Class """
         if type(self._house) is str:
             data = self._client.storage['houses'].get(self._house)
             if data:
@@ -87,6 +118,7 @@ class UserTyping(DataClassObject):
 
     @property
     def room(self) -> Optional[TextRoom]:
+        """ Room object of the Context Class """
         if type(self._room) is str:
             data = self._client.storage['rooms']['house'].get(self._room)
             if data:
@@ -102,12 +134,20 @@ class UserTyping(DataClassObject):
 
     @property
     def author_id(self) -> Optional[str]:
+        """ ID of the parent Author object of the Context Class """
         return getattr(self, '_author_id', None)
 
     @property
     def house_id(self) -> Optional[str]:
+        """ ID of the parent House object of the Context Class """
         return getattr(self, '_house_id', None)
 
     @property
+    def is_house_typing(self) -> bool:
+        """ Returns whether the typing is inside a house """
+        return self.house_id is not None
+
+    @property
     def room_id(self) -> Optional[str]:
+        """ ID of the parent Room object of the Context Class """
         return getattr(self, '_room_id', None)